refactor(package-manager): move trusted-keys to the shared surface
trusted-keys.ts is a platform-agnostic trust anchor (keyId -> public key). Moving it to backend/shared/utils/vpp puts it on the byte-identical shared surface, so the editor and openplc-web are guaranteed (via ci-sync) to trust the same signing keys. package-manager-module.ts stays editor-only — it is an Electron adapter (app.getPath, local fs, registry) and the web will provide its own under backend/web, reusing the shared verifier + trusted keys.
